+++ src/main/java/com/genius/gitget/store/item/facade/StoreFacadeService.java
@@ -1,5 +1,11 @@
 package com.genius.gitget.store.item.facade;
 
+import com.genius.gitget.challenge.certification.dto.CertificationRequest;
+import com.genius.gitget.challenge.certification.service.CertificationService;
+import com.genius.gitget.challenge.myChallenge.dto.ActivatedResponse;
+import com.genius.gitget.challenge.myChallenge.dto.DoneResponse;
+import com.genius.gitget.challenge.myChallenge.dto.RewardRequest;
+import com.genius.gitget.challenge.myChallenge.service.MyChallengeService;
 import com.genius.gitget.challenge.user.domain.User;
 import com.genius.gitget.challenge.user.service.UserService;
 import com.genius.gitget.global.util.exception.BusinessException;
@@ -19,14 +25,21 @@ import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
-//@Service
-//@Transactional(readOnly = true)
+@Service
+@Transactional(readOnly = true)
 @RequiredArgsConstructor
-public class StoreFacadeImpl implements StoreFacade {
-    private final UserService userService;
+public class StoreFacadeService implements StoreFacade {
     private final ItemService itemService;
     private final OrdersService ordersService;
+
+    private final UserService userService;
+    private final CertificationService certificationService;
+
+    //TODO: MyChallengeService 말고 책임이 분명한 Service를 만들어서 적용하기
+    private final MyChallengeService myChallengeService;
 
     private final PaymentRepository paymentRepository;
 
@@ -70,7 +83,25 @@ public class StoreFacadeImpl implements StoreFacade {
 
     @Override
     public ItemUseResponse useItem(User user, Long itemId, Long instanceId, LocalDate currentDate) {
-        return null;
+        Item item = itemService.findById(itemId);
+        Orders orders = ordersService.findByOrderInfo(user.getId(), itemId);
+
+        if (!orders.hasItem()) {
+            throw new BusinessException(ErrorCode.HAS_NO_ITEM);
+        }
+
+        switch (item.getItemCategory()) {
+            case PROFILE_FRAME -> {
+                return useFrameItem(user.getId(), orders);
+            }
+            case CERTIFICATION_PASSER -> {
+                return usePasserItem(orders, instanceId, currentDate);
+            }
+            case POINT_MULTIPLIER -> {
+                return useMultiplierItem(orders, instanceId, currentDate);
+            }
+        }
+        throw new BusinessException(ErrorCode.ORDERS_NOT_FOUND);
     }
 
     @Override
@@ -102,7 +133,14 @@ public class StoreFacadeImpl implements StoreFacade {
      */
     @Override
     public ItemUseResponse usePasserItem(Orders orders, Long instanceId, LocalDate currentDate) {
-        return null;
+        Long userId = orders.getUser().getId();
+        Long itemId = orders.getItem().getId();
+        ActivatedResponse activatedResponse = certificationService.passCertification(
+                userId,
+                new CertificationRequest(instanceId, currentDate));
+        activatedResponse.setItemId(itemId);
+        ordersService.useItem(orders);
+        return activatedResponse;
     }
 
     /**
@@ -114,14 +152,35 @@ public class StoreFacadeImpl implements StoreFacade {
      */
     @Override
     public ItemUseResponse useMultiplierItem(Orders orders, Long instanceId, LocalDate currentDate) {
-        return null;
+        User user = orders.getUser();
+        DoneResponse doneResponse = myChallengeService.getRewards(
+                new RewardRequest(user, instanceId, currentDate), true
+        );
+        doneResponse.setItemId(orders.getItem().getId());
+        ordersService.useItem(orders);
+        return doneResponse;
     }
 
-    /**
-     *
-     */
     @Override
     public List<ProfileResponse> unmountFrame(User user) {
-        return null;
+        List<ProfileResponse> profileResponses = new ArrayList<>();
+        List<Orders> frameOrders = ordersService.findAllUsingFrames(user.getId());
+
+        for (Orders frameOrder : frameOrders) {
+            validateUnmountCondition(frameOrder);
+            frameOrder.updateEquipStatus(EquipStatus.AVAILABLE);
+            profileResponses.add(ProfileResponse.createByEntity(frameOrder));
+        }
+
+        return profileResponses;
+    }
+
+    private void validateUnmountCondition(Orders orders) {
+        if (orders.getItem().getItemCategory() != ItemCategory.PROFILE_FRAME) {
+            throw new BusinessException(ErrorCode.ITEM_NOT_FOUND);
+        }
+        if (orders.getEquipStatus() != EquipStatus.IN_USE) {
+            throw new BusinessException(ErrorCode.IN_USE_FRAME_NOT_FOUND);
+        }
     }
 }
